stock/dynamic_fetch/fetching_post_url_gcloud/requirements.txt
google-cloud-functions
praw
requests
python-dotenv
pytz
flask
beautifulsoup4
pymongo
dnspython
webdriver-manager
selenium
concurrent.futures
